changed rangeCount() return to long
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@494 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/inf/ClosureStats.java
@@ -43,7 +43,7 @@ public class ClosureStats {
      * change in the #of statements in the database across the closure
      * operation.
      */
-    public int nentailments;
+    public long nentailments;
 
     /**
      * Time to compute the entailments and store them within the database
